refactor: update package name from blablabla to api in BookSearch and BookService
