+++ src/main/java/app/main/LibraryApp/api/BookSearch.java
@@ -1,4 +1,4 @@
-package app.main.LibraryApp.blablabla;
+package app.main.LibraryApp.api;
 
 import java.io.IOException;
 import java.net.URI;
